Замена double на long double

+++ 01_week/tasks/length_lit/length_lit.cpp
@@ -7,62 +7,62 @@ constexpr double INCHES_IN_FOOT = 12.0;
 constexpr double CM_IN_INCH = 2.54;
 
 // Метры в футы
-constexpr double operator"" _m_to_ft(double meters) {
-    return static_cast<double>(meters / METERS_IN_FOOT);
+constexpr double operator"" _m_to_ft(long double meters) {
+    return static_cast<long double>(meters / METERS_IN_FOOT);
 }
 
 // Метры в дюймы
-constexpr double operator"" _m_to_in(double meters) {
+constexpr double operator"" _m_to_in(long double meters) {
     return static_cast<double>(meters * CM_IN_METER / CM_IN_INCH);
 }
 
 // Метры в сантиметры
-constexpr double operator"" _m_to_cm(double meters) {
+constexpr double operator"" _m_to_cm(long double meters) {
     return static_cast<double>(meters * CM_IN_METER);
 }
 
 // Сантиметры в метры
-constexpr double operator"" _cm_to_m(double centimeters) {
-    return static_cast<double>(centimeters / CM_IN_METER);
+constexpr double operator"" _cm_to_m(long double centimeters) {
+    return static_cast<long double>(centimeters / CM_IN_METER);
 }
 
 // Сантиметры в футы
-constexpr double operator"" _cm_to_ft(double centimeters) {
+constexpr double operator"" _cm_to_ft(long double centimeters) {
     return static_cast<double>(centimeters / CM_IN_METER / METERS_IN_FOOT);
 }
 
 // Сантиметры в дюймы
-constexpr double operator"" _cm_to_in(double centimeters) {
+constexpr double operator"" _cm_to_in(long double centimeters) {
     return static_cast<double>(centimeters / CM_IN_INCH);
 }
 
 // Футы в метры
-constexpr double operator"" _ft_to_m(double feet) {
+constexpr double operator"" _ft_to_m(long double feet) {
     return static_cast<double>(feet * METERS_IN_FOOT);
 }
 
 // Футы в сантиметры
-constexpr double operator"" _ft_to_cm(double feet) {
+constexpr double operator"" _ft_to_cm(long double feet) {
     return static_cast<double>(feet * METERS_IN_FOOT * CM_IN_METER);
 }
 
 // Футы в дюймы
-constexpr double operator"" _ft_to_in(double feet) {
+constexpr double operator"" _ft_to_in(long double feet) {
     return static_cast<double>(feet * INCHES_IN_FOOT);
 }
 
 
 // Дюймы в метры
-constexpr double operator"" _in_to_m(double inches) {
+constexpr double operator"" _in_to_m(long double inches) {
     return static_cast<double>(inches * CM_IN_INCH / CM_IN_METER);
 }
 
 // Дюймы в сантиметры
-constexpr double operator"" _in_to_cm(double inches) {
+constexpr double operator"" _in_to_cm(long double inches) {
     return static_cast<double>(inches * CM_IN_INCH);
 }
 
 // Дюймы в футы
-constexpr double operator"" _in_to_ft(double inches) {
+constexpr double operator"" _in_to_ft(long double inches) {
     return static_cast<double>(inches / INCHES_IN_FOOT);
 }
